modules/graph-canvas-compiler/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "graph-canvas-compiler"
version = "0.1.0"
description = "Compiler for graph canvas definitions"
requires-python = ">=3.11"
dependencies = [
    "ruamel.yaml>=0.18.0",
]

[tool.hatch.build.targets.wheel]
packages = ["amplifier_module_graph_canvas_compiler"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
